添加拖拽类型接口

+++ window_scene/interfaces/kits/napi/scene_session_manager/js_scene_utils.cpp
@@ -1342,8 +1342,7 @@ napi_value CreateJsSessionEventParam(napi_env env, const SessionEventParam& para
     napi_set_named_property(env, objValue, "pointerY", CreateJsValue(env, param.pointerY_));
     napi_set_named_property(env, objValue, "sessionWidth", CreateJsValue(env, param.sessionWidth_));
     napi_set_named_property(env, objValue, "sessionHeight", CreateJsValue(env, param.sessionHeight_));
-    napi_set_named_property(env, objValue, "dragResizeType",
-        CreateJsValue(env, static_cast<uint32_t>(param.dragResizeType)));
+    napi_set_named_property(env, objValue, "dragResizeType", CreateJsValue(env, param.dragResizeType));
     return objValue;
 }
